404 error page now need to be on it's own and static
